use cmake to build cpp file

+++ app/src/main/java/com/example/rmd2k/guitarstudio_android/Tools/Tuner/FFT.java
@@ -5,5 +5,5 @@ public class FFT {
         System.loadLibrary("FFT");
     }
 
-    public static native byte[] fft(short[] fftData);
+    public static native short[] fft(short[] fftData);
 }
