+++ app/src/main/java/com/example/rmd2k/guitarstudio_android/Tools/Tuner/TunerActivity.java
@@ -110,30 +110,10 @@ break;
                 audioRecord.startRecording();
                 isRecording = true;
 
-//                int audioTrackMinBufferSize = AudioTrack.getMinBufferSize(44100, AudioFormat.CHANNEL_OUT_MONO, AudioFormat.ENCODING_PCM_16BIT);
-//                final AudioTrack audioTrack = new AudioTrack(AudioManager.STREAM_MUSIC, 44100, AudioFormat.CHANNEL_OUT_MONO, AudioFormat.ENCODING_PCM_16BIT, audioTrackMinBufferSize, AudioTrack.MODE_STREAM);
-//                audioTrack.play();
-
-//                final int maxCR = Visualizer.getMaxCaptureRate();
-//                Visualizer visualizer = new Visualizer(audioTrack.getAudioSessionId());
-//                visualizer.setCaptureSize(Visualizer.getCaptureSizeRange()[1]);
-//                visualizer.setDataCaptureListener(new Visualizer.OnDataCaptureListener() {
-//                    @Override
-//                    public void onWaveFormDataCapture(Visualizer visualizer, byte[] waveform, int samplingRate) {
-//                        //Log.e("KONG", "aaa");
-//                    }
-//
-//                    @Override
-//                    public void onFftDataCapture(Visualizer visualizer, byte[] fft, int samplingRate) {
-//                        //Log.e("KONG", "aaa:" + fft.length);
-//                    }
-//                }, maxCR / 2, true, true);
-//                visualizer.setEnabled(true);
-
                 while (isRecording) {
                     int bufferReadReault = audioRecord.read(buffer, 0, audioRecordMinBufferSize);
                     Log.e("KONG", "count:" + bufferReadReault);
-                    byte[] fft = FFT.fft(buffer);
+                    short[] fft = FFT.fft(buffer);
 
 //                    for (int i = 0; i < bufferReadReault; i++) {
 //                        Log.i("KONG", String.valueOf(buffer[i]));
@@ -147,28 +127,11 @@ break;
                     instrumentPanelView.invalidate();
                     //audioTrack.write(buffer, 0, bufferReadReault);
 
-
                 }
                 audioRecord.stop();
-                //audioTrack.stop();
-                //visualizer.setEnabled(false);
             }
         }).start();
     }
-
-//    private short[] fft(short[] data, int count) {
-//        short[] fft = new short[512];
-//        for (int hz = 0; hz < 512; hz++) {
-//            fft[hz] = 0;
-//            for (int i = 0; i < count; i++) {
-//                fft[hz] += data[i] * Math.exp(2 * Math.PI * Math.sqrt(-1) * hz);
-//                Log.e("KONG", "fft:" + data[i]);
-//            }
-//            fft[hz] /= count;
-//        }
-//
-//        return fft;
-//    }
 
     @Override
     protected void onDestroy() {
